fix userDefaults
