+++ Recepies/Modules/Authorization/Model/AuthService.swift
@@ -58,83 +58,3 @@ final class AuthService: AuthServiceProtocol {
         return (isFormatOk, isValid)
     }
 }
-
-// протокол мементор
-// структру подписана должна быть
-struct User: Codable {
-    var login: String
-    var password: String
-    var nickName: String
-    var avatar: String
-}
-
-// 1. Originator
-
-// каретакер
-final class Caretaker {
-    static var shared = Caretaker()
-    private let encoder = JSONEncoder()
-    private let decoder = JSONDecoder()
-    private let key = "records"
-
-    private var user: User = .init(login: "", password: "", nickName: "", avatar: "avatar") {
-        didSet {
-            save(records: user)
-        }
-    }
-
-    func updateLogin(login: String) {
-        user.login = login
-    }
-
-    func updatePassword(password: String) {
-        user.password = password
-    }
-
-    func updateNameUser(name: String) {
-        user.nickName = name
-    }
-
-    func loadUser() -> User {
-        user = load() ?? User(login: "", password: "", nickName: "", avatar: "")
-        return user
-    }
-
-    func saveImage(data: Data) {
-        let encoded = try? PropertyListEncoder().encode(data)
-        UserDefaults.standard.set(encoded, forKey: "avatar")
-    }
-
-    func loadImage() -> Data? {
-        guard let data = UserDefaults.standard.data(forKey: "avatar") else {
-            return nil
-        }
-        do {
-            return try PropertyListDecoder().decode(Data.self, from: data)
-        } catch {
-            print(error)
-            return nil
-        }
-    }
-
-    private func save(records: User) {
-        do {
-            let data = try encoder.encode(records)
-            UserDefaults.standard.set(data, forKey: key)
-        } catch {
-            print(error)
-        }
-    }
-
-    private func load() -> User? {
-        guard let data = UserDefaults.standard.data(forKey: key) else {
-            return nil
-        }
-        do {
-            return try decoder.decode(User.self, from: data)
-        } catch {
-            print(error)
-            return nil
-        }
-    }
-}
