+++ Recepies/Modules/Authorization/Model/User.swift
@@ -1,0 +1,14 @@
+// User.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// User data
+struct User: Codable {
+    /// Login
+    var login: String
+    /// Password
+    var password: String
+    /// First and last name
+    var nickName: String
+    /// User avatar
+    var avatar: String
+}
